Road accidents analysis using MS Excel
In this project, I used MS Excel for data cleaning and data presentation This whole project is about road accidents. I picked up the data from a random website. First, I cleaned the data and removed any abnormalities I found in the data. In the second step, I used the cleaned data to gather the information I needed for the project. The usage of a pivot table proved useful in this phase. In the third phase, I started building the interactive dashboard using the data from the second step. To conclude, I can say that by doing this project, I learned how to clean data, group data, and finally, how to present it in the form of an interactive dashboard.